Use GLFW instead of SDL2
It is much lighter and simpler just for our purposes of GL window context creation.
